requirement.txt
Django==3.1.3
django-ckeditor==6.0.0
Pillow==8.0.1
psycopg2==2.8.6psycopg2
